// frontend/src/api/models/SuggestionOut.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overmind Core
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SuggestionOut
 */
export interface SuggestionOut {
    /**
     * 
     * @type {string}
     * @memberof SuggestionOut
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof SuggestionOut
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof SuggestionOut
     */
    description: string;
    /**
     * 
     * @type {number}
     * @memberof SuggestionOut
     */
    newPromptVersion?: number | null;
    /**
     * 
     * @type {string}
     * @memberof SuggestionOut
     */
    newPromptText?: string | null;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof SuggestionOut
     */
    scores?: { [key: string]: any; } | null;
    /**
     * 
     * @type {string}
     * @memberof SuggestionOut
     */
    status?: string;
    /**
     * 
     * @type {number}
     * @memberof SuggestionOut
     */
    vote?: number;
    /**
     * 
     * @type {string}
     * @memberof SuggestionOut
     */
    feedback?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SuggestionOut
     */
    createdAt?: string | null;
}

/**
 * Check if a given object implements the SuggestionOut interface.
 */
export function instanceOfSuggestionOut(value: object): value is SuggestionOut {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    return true;
}

export function SuggestionOutFromJSON(json: any): SuggestionOut {
    return SuggestionOutFromJSONTyped(json, false);
}

export function SuggestionOutFromJSONTyped(json: any, ignoreDiscriminator: boolean): SuggestionOut {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'title': json['title'],
        'description': json['description'],
        'newPromptVersion': json['new_prompt_version'] == null ? undefined : json['new_prompt_version'],
        'newPromptText': json['new_prompt_text'] == null ? undefined : json['new_prompt_text'],
        'scores': json['scores'] == null ? undefined : json['scores'],
        'status': json['status'] == null ? undefined : json['status'],
        'vote': json['vote'] == null ? undefined : json['vote'],
        'feedback': json['feedback'] == null ? undefined : json['feedback'],
        'createdAt': json['created_at'] == null ? undefined : json['created_at'],
    };
}

export function SuggestionOutToJSON(json: any): SuggestionOut {
    return SuggestionOutToJSONTyped(json, false);
}

export function SuggestionOutToJSONTyped(value?: SuggestionOut | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'description': value['description'],
        'new_prompt_version': value['newPromptVersion'],
        'new_prompt_text': value['newPromptText'],
        'scores': value['scores'],
        'status': value['status'],
        'vote': value['vote'],
        'feedback': value['feedback'],
        'created_at': value['createdAt'],
    };
}